+++ packages/remix-testing/CHANGELOG.md
@@ -1,5 +1,29 @@
 # `@remix-run/testing`
 
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@2.7.2`
+  - `@remix-run/react@2.7.2`
+
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@2.7.1`
+  - `@remix-run/react@2.7.1`
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/react@2.7.0`
+  - `@remix-run/node@2.7.0`
+
 ## 2.6.0
 
 ### Patch Changes